+++ packages/bmg/src/support/toPredicateFunc.ts
@@ -1,8 +1,11 @@
+import { isPredicate, evaluate } from '@enspirit/predicate';
 import { Predicate, PredicateFunc, Tuple } from '../types';
 
 export const toPredicateFunc = (p: Predicate): PredicateFunc => {
   if (typeof(p) === 'function') {
     return p as PredicateFunc;
+  } else if (isPredicate(p)) {
+    return (t: Tuple) => evaluate(p, t as Record<string, unknown>);
   } else {
     const expected = p as Tuple;
     return (t: Tuple) => {